price-anomaly/pyproject.toml
[tool.poetry]
name = "price-anomaly"
version = "0.1.0"
description = "Challenge MLE Mercado Libre 2024"
authors = ["Pedro <48163181+neuralgremlin@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
scikit-learn = "^1.5.0"
robyn = "^0.56.0"
pandas = "^2.2.2"
numpy = "^2.0.0"
prometheus-client = "^0.20.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
plotly = "^5.22.0"
seaborn = "^0.13.2"
ruamel-yaml = "^0.18.6"
python-box = {version = "^5.0.0", extras = ["ruamel.yaml"]}
tqdm = "^4.66.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
